# docs/openapi.courtly.yaml
openapi: 3.1.0
info:
  title: Courtly API
  version: 1.0.0
  description: >
    Full-scope API for Courtly tennis court booking platform.
    Includes auth, courts, day/week availability, hold-based booking,
    user cabinet, moderator/admin dashboard, dynamic RBAC/policy management,
    and event log operations.
servers:
  - url: https://api.courtly.local
    description: Production
  - url: http://localhost:8000
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/courts:
    get:
      tags: [Courts]
      summary: List courts with filters
      security: []
      parameters:
        - in: query
          name: city
          schema: { type: string }
        - in: query
          name: district
          schema: { type: string }
        - in: query
          name: lat
          schema: { type: number, format: float }
        - in: query
          name: lng
          schema: { type: number, format: float }
        - in: query
          name: radius_km
          schema: { type: number, format: float, default: 10 }
      responses:
        "200":
          description: Court list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/CourtSummary"
    post:
      tags: [Courts]
      summary: Create court
      x-rbac:
        permission: courts:create
        ownership: none
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourtCreateRequest"
      responses:
        "201":
          description: Court created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourtDetail"
        "403":
          description: Forbidden
  /api/courts/{court_id}:
    get:
      tags: [Courts]
      summary: Get court detail
      security: []
      parameters:
        - $ref: "#/components/parameters/CourtId"
      responses:
        "200":
          description: Court detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourtDetail"
        "404":
          description: Court not found
    patch:
      tags: [Courts]
      summary: Update court
      x-rbac:
        permission: courts:update
        ownership: own_court_or_admin
      parameters:
        - $ref: "#/components/parameters/CourtId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourtUpdateRequest"
      responses:
        "200":
          description: Court updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourtDetail"
        "403":
          description: Forbidden
    delete:
      tags: [Courts]
      summary: Delete court
      x-rbac:
        permission: courts:delete
        ownership: own_court_or_admin
      parameters:
        - $ref: "#/components/parameters/CourtId"
      responses:
        "204":
          description: Court deleted
        "403":
          description: Forbidden
  /api/courts/{court_id}/availability:
    get:
      tags: [Availability]
      summary: Get court availability in day or week mode
      security: []
      parameters:
        - $ref: "#/components/parameters/CourtId"
        - in: query
          name: mode
          required: true
          schema:
            type: string
            enum: [day, week]
        - in: query
          name: date
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Availability data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilityResponse"
  /api/bookings/hold:
    post:
      tags: [Bookings]
      summary: Hold slots temporarily
      x-rbac:
        permission: bookings:hold
        ownership: principal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingHoldRequest"
      responses:
        "201":
          description: Hold created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingHoldResponse"
        "409":
          description: Slot conflict
        "422":
          description: Invalid slot selection
  /api/bookings/confirm:
    post:
      tags: [Bookings]
      summary: Confirm booking from hold token
      x-rbac:
        permission: bookings:confirm
        ownership: principal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingConfirmRequest"
      responses:
        "201":
          description: Booking confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResponse"
        "409":
          description: Hold expired or conflicted
  /api/bookings/{booking_id}/cancel:
    post:
      tags: [Bookings]
      summary: Cancel booking
      x-rbac:
        permission: bookings:cancel
        ownership: own_booking_or_admin
      parameters:
        - $ref: "#/components/parameters/BookingId"
      responses:
        "200":
          description: Booking cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResponse"
        "403":
          description: Forbidden
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CourtId:
      in: path
      name: court_id
      required: true
      schema:
        type: string
        format: uuid
    BookingId:
      in: path
      name: booking_id
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    CourtCreateRequest:
      type: object
      required: [name, city, district, address, timezone]
      properties:
        name: { type: string }
        city: { type: string }
        district: { type: string }
        address: { type: string }
        timezone: { type: string, example: Europe/Kyiv }
        lat: { type: number, format: float }
        lng: { type: number, format: float }
        indoor: { type: boolean, default: false }
        surface: { type: string, example: hard }
        price_per_hour: { type: number, format: float, minimum: 0 }
        working_hours:
          type: array
          items:
            $ref: "#/components/schemas/WorkingHours"
    CourtUpdateRequest:
      type: object
      additionalProperties: false
      properties:
        name: { type: string }
        city: { type: string }
        district: { type: string }
        address: { type: string }
        indoor: { type: boolean }
        surface: { type: string }
        price_per_hour: { type: number, format: float, minimum: 0 }
    CourtSummary:
      type: object
      required: [id, name, city, district, next_available_slot]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        city: { type: string }
        district: { type: string }
        address: { type: string }
        rating: { type: number, format: float }
        distance_km: { type: number, format: float }
        price_from: { type: number, format: float }
        next_available_slot: { type: string, format: date-time }
    CourtDetail:
      allOf:
        - $ref: "#/components/schemas/CourtSummary"
        - type: object
          properties:
            description: { type: string }
            indoor: { type: boolean }
            surface: { type: string }
            media_urls:
              type: array
              items: { type: string, format: uri }
            working_hours:
              type: array
              items:
                $ref: "#/components/schemas/WorkingHours"
    WorkingHours:
      type: object
      required: [weekday, open_time, close_time]
      properties:
        weekday:
          type: integer
          minimum: 0
          maximum: 6
        open_time:
          type: string
          pattern: "^[0-2][0-9]:[0-5][0-9]$"
        close_time:
          type: string
          pattern: "^[0-2][0-9]:[0-5][0-9]$"
    AvailabilitySlot:
      type: object
      required: [start_at, end_at, state]
      properties:
        start_at: { type: string, format: date-time }
        end_at: { type: string, format: date-time }
        state:
          type: string
          enum: [free, held, booked, disabled, past]
        held_until:
          type: string
          format: date-time
          nullable: true
    AvailabilityDay:
      type: object
      required: [date, slots]
      properties:
        date: { type: string, format: date }
        slots:
          type: array
          items:
            $ref: "#/components/schemas/AvailabilitySlot"
    AvailabilityResponse:
      type: object
      required: [mode, days]
      properties:
        mode:
          type: string
          enum: [day, week]
        days:
          type: array
          items:
            $ref: "#/components/schemas/AvailabilityDay"
    BookingHoldRequest:
      type: object
      required: [court_id, slot_starts]
      properties:
        court_id:
          type: string
          format: uuid
        slot_starts:
          type: array
          minItems: 2
          items:
            type: string
            format: date-time
      description: "slot_starts must represent contiguous 30-minute slots, minimum total 60 minutes."
    BookingHoldResponse:
      type: object
      required: [hold_token, held_until, court_id, slot_starts]
      properties:
        hold_token: { type: string }
        held_until: { type: string, format: date-time }
        court_id: { type: string, format: uuid }
        slot_starts:
          type: array
          items: { type: string, format: date-time }
    BookingConfirmRequest:
      type: object
      required: [hold_token]
      properties:
        hold_token: { type: string }
    BookingResponse:
      type: object
      required: [id, court_id, user_id, status, start_at, end_at, total_price]
      properties:
        id: { type: string, format: uuid }
        court_id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        status:
          type: string
          enum: [confirmed, active, completed, cancelled]
        start_at: { type: string, format: date-time }
        end_at: { type: string, format: date-time }
        total_price: { type: number, format: float }
        created_at: { type: string, format: date-time }
...